VM/ASM: partial test of comparisons
Currently stuck with an issue considering implicit integer conversions:
  0000:u16 < fffe:s16
is false, because both operands are considered as signed in this case.
It implies a deep review of the equalization/normalization of numbers
=> WIP.

+++ misc/cmptest16.c
@@ -1,7 +1,6 @@
 #include <stdio.h>
-#include <stdint.h>
 
-uint16_t values[] = {
+int values[] = {
     0x0000,
     0x0001,
     0x00fe,
@@ -18,16 +17,16 @@ uint16_t values[] = {
 
 int main() {
     unsigned int i, j;
-    for (i = 0; i < sizeof(values) / sizeof(uint16_t); i++) {
-        uint16_t ux = values[i];
-        int16_t  sx = values[i];
-        for (j = 0; j < sizeof(values) / sizeof(uint16_t); j++) {
-            uint16_t uy = values[j];
+    for (i = 0; i < sizeof(values) / sizeof(int); i++) {
+        unsigned short ux = values[i];
+        short  sx = values[i];
+        for (j = 0; j < sizeof(values) / sizeof(int); j++) {
+            unsigned short uy = values[j];
             printf("\"%04x u16 %04x u16 %u %u %u %u %u %u\",\n",
                     ux & 0xffff, uy & 0xffff, ux == uy, ux != uy, ux < uy, ux <= uy, ux > uy, ux >= uy);
             printf("\"%04x s16 %04x u16 %u %u %u %u %u %u\",\n",
                     sx & 0xffff, uy & 0xffff, sx == uy, sx != uy, sx < uy, sx <= uy, sx > uy, sx >= uy);
-            int16_t  sy = values[j];
+            short  sy = values[j];
             printf("\"%04x u16 %04x s16 %u %u %u %u %u %u\",\n",
                     ux & 0xffff, sy & 0xffff, ux == sy, ux != sy, ux < sy, ux <= sy, ux > sy, ux >= sy);
             printf("\"%04x s16 %04x s16 %u %u %u %u %u %u\",\n",
